requirements.txt
# ─── Deep Learning Framework ─────────────────────────────────────────────────
torch>=2.2.0                # PyTorch – backbone for all YOLO models
torchvision>=0.17.0         # vision utilities

# ─── Object Detection ────────────────────────────────────────────────────────
ultralytics>=8.2.0          # YOLOv8n – vehicle & plate detection

# ─── OCR ─────────────────────────────────────────────────────────────────────
easyocr>=1.7.1              # alphanumeric plate OCR (no training needed)

# ─── Image Processing ────────────────────────────────────────────────────────
opencv-python>=4.9.0        # preprocessing, cropping, bounding boxes
Pillow>=10.3.0              # image I/O utility

# ─── Numerical / Data ────────────────────────────────────────────────────────
numpy>=1.26.0
pandas>=2.2.0

# ─── Metrics & Evaluation ────────────────────────────────────────────────────
scikit-learn>=1.4.0

# ─── Web Interface (Streamlit) ───────────────────────────────────────────────
streamlit>=1.35.0           # UI / deployment – no frontend code needed

# ─── Model Optimization (ONNX + Quantization) ───────────────────────────────
onnx>=1.16.0                # ONNX model export
onnxruntime>=1.18.0         # CPU inference via ONNX
onnxruntime-gpu>=1.18.0     # GPU inference (optional – comment out if no GPU)

# ─── Utilities ───────────────────────────────────────────────────────────────
pyyaml>=6.0.1
tqdm>=4.66.0
matplotlib>=3.8.0
seaborn>=0.13.0
requests>=2.31.0
python-dotenv>=1.0.0
